Add files via upload
new function based on Houria Madani's MATLAB equivalent code

This model_bkgnd function fits the background of the given frame, 
ignoring bright spots (cities), as a polynomial with a step in the 
middle of the frame. We have substituted the numpy.polyfit and 
numpy.polyval fitting functions for a long-form linear algebra 
least-squares fit to account for the step that interrupts the 
polynomial fit.